# src/main/java/com/preferences/domain/dto/PreferencesRequest.java
package com.preferences.domain.dto;

import jakarta.validation.constraints.NotNull;
import com.fasterxml.jackson.annotation.JsonProperty;

public class PreferencesRequest {
    
    @NotNull(message = "emailNotifications is required")
    @JsonProperty("emailNotifications")
    private Boolean emailNotifications;
    
    @NotNull(message = "smsNotifications is required")
    @JsonProperty("smsNotifications")
    private Boolean smsNotifications;
    
    @NotNull(message = "language is required")
    @JsonProperty("language")
    private String language;
    
    @JsonProperty("timezone")
    private String timezone;
    
    @JsonProperty("marketingConsent")
    private Boolean marketingConsent;

    public PreferencesRequest() {
    }

    public PreferencesRequest(Boolean emailNotifications, Boolean smsNotifications, String language, 
                             String timezone, Boolean marketingConsent) {
        this.emailNotifications = emailNotifications;
        this.smsNotifications = smsNotifications;
        this.language = language;
        this.timezone = timezone;
        this.marketingConsent = marketingConsent;
    }

    public Boolean getEmailNotifications() {
        return emailNotifications;
    }

    public void setEmailNotifications(Boolean emailNotifications) {
        this.emailNotifications = emailNotifications;
    }

    public Boolean getSmsNotifications() {
        return smsNotifications;
    }

    public void setSmsNotifications(Boolean smsNotifications) {
        this.smsNotifications = smsNotifications;
    }

    public String getLanguage() {
        return language;
    }

    public void setLanguage(String language) {
        this.language = language;
    }

    public String getTimezone() {
        return timezone;
    }

    public void setTimezone(String timezone) {
        this.timezone = timezone;
    }

    public Boolean getMarketingConsent() {
        return marketingConsent;
    }

    public void setMarketingConsent(Boolean marketingConsent) {
        this.marketingConsent = marketingConsent;
    }
}